Utilizando Add-Migrations em SisClientes

+++ SisProdutos/Data/Dtos/Produto/ReadProdutoDto.cs
@@ -9,7 +9,7 @@ namespace SisProdutos
         public int Id { get; set; }
         public string Nome { get; set; }
         public string Descicao { get; set; }
-        public decimal Preco { get; set; }
+        public float Preco { get; set; }
         public string PalavraChave { get; set; }
         public string Categoria { get; set; }
         public int CidadeId { get; set; }
